+++ presentation/src/main/java/com/nexters/pimo/ui/component/FimoPostGrid.kt
@@ -1,0 +1,39 @@
+package com.nexters.pimo.ui.component
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import com.nexters.pimo.domain.model.Post
+
+@Composable
+fun FimoPostGrid(
+    modifier: Modifier = Modifier,
+    posts: List<Post>,
+    onPostClick: (Post) -> Unit
+) {
+    LazyVerticalGrid(
+        modifier = modifier,
+        columns = GridCells.Fixed(2),
+        verticalArrangement = Arrangement.spacedBy(4.dp),
+        horizontalArrangement = Arrangement.spacedBy(4.dp)
+    ) {
+        items(posts) {
+            AsyncImage(
+                model = it.textImages[0].imageUrl,
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+                modifier = Modifier
+                    .aspectRatio(1f)
+                    .clickable { onPostClick(it) }
+            )
+        }
+    }
+}
